implementation of status change

+++ src/modules/posts/application/use-cases/delete-post.use-case.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { LoggingService } from '../../../shared/logging/domain/services/logging.service';
 import { PostRepository } from '../../domain/repositories/post.repository';
+import { UserEntity } from 'src/modules/users/domain/entities/user.entity';
 
 @Injectable()
 export class DeletePostUseCase {
@@ -9,8 +10,13 @@ export class DeletePostUseCase {
     private readonly loggingService: LoggingService,
   ) {}
 
-  public async execute(id: string): Promise<void> {
-    this.loggingService.log('DeletePostUseCase.execute');
-    await this.postRepository.deletePost(id);
-  }
+  public async execute(id: string, currentUser: UserEntity): Promise<void> {
+    const post = await this.postRepository.getPostById(id);
+    if (post){
+      if (currentUser.getRole() == 'admin' || currentUser.id === post.authorId) {
+        this.loggingService.log('DeletePostUseCase.execute');
+        await this.postRepository.deletePost(id);
+      }
+    }
+  }  
 }
